+++ tailwind/tailwind.settings.aspectRatio.js
@@ -1,8 +1,0 @@
-/*
- * Tailwind Screens Settings
- */
-
-module.exports = {
-  landscape: '3 / 2',
-  portrait: '2 / 3',
-};