+++ tailwind/tailwindcss-theming/src/utils/createObject.js
@@ -1,9 +1,0 @@
-module.exports = (obj, colorName, colorLevel, value) => {
-  if (obj[colorName]) {
-    obj[colorName][colorLevel] = value;
-  } else {
-    obj[colorName] = {
-      [colorLevel]: value,
-    };
-  }
-};